requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
python-multipart
requests
starlette

## uvicorn main2:app --host 0.0.0.0 --port 8000
## ngrok http 8000
## ngrok authtoken 2w8sGYxmwfHWPTa8gmxjMowA0hs_2hbiTHuXRdSLeD4mTxGyS
